VIH-4718: Added new endpoint
