+++ .gitignore
@@ -139,6 +139,10 @@ _TeamCity*
 _NCrunch_*
 .*crunch*.local.xml
 nCrunchTemp_*
+*.ncrunch*
+*ncrunchproject
+*ncrunchsolution
+*crunchsolution.cache
 
 # MightyMoose
 *.mm.*
