+++ BookingsAPI/Bookings.IntegrationTests/Database/Queries/GetPersonByUsernameQueryHandlerTests.cs
@@ -1,9 +1,7 @@
-using System;
 using System.Linq;
 using System.Threading.Tasks;
 using Bookings.DAL;
 using Bookings.DAL.Queries;
-using Faker;
 using FluentAssertions;
 using NUnit.Framework;
